docs: general purpose readme
Adds a general purpose readme to the codebase. This may not have all the details we need yet, but it's a good start. Most of the language used is abstract enough for most contexts, but should be modified to fit the specific use case once this code instantiated from the Github template.